EDU-686 Prevent public rooms and documents deletion

+++ src/server/room-controller.js
@@ -5,7 +5,6 @@ import urlUtils from '../utils/url-utils.js';
 import PageRenderer from './page-renderer.js';
 import { PAGE_NAME } from '../domain/page-name.js';
 import permissions from '../domain/permissions.js';
-import { ROOM_ACCESS } from '../domain/constants.js';
 import requestUtils from '../utils/request-utils.js';
 import RoomService from '../services/room-service.js';
 import UserService from '../services/user-service.js';
@@ -16,6 +15,7 @@ import DocumentService from '../services/document-service.js';
 import needsPermission from '../domain/needs-permission-middleware.js';
 import ClientDataMappingService from '../services/client-data-mapping-service.js';
 import { validateBody, validateParams, validateQuery } from '../domain/validation-middleware.js';
+import { ROOM_ACCESS, NOT_ROOM_OWNER_ERROR_MESSAGE, NOT_ROOM_OWNER_OR_MEMBER_ERROR_MESSAGE } from '../domain/constants.js';
 import {
   postRoomBodySchema,
   getRoomParamsSchema,
@@ -82,7 +82,7 @@ export default class RoomController {
     }
 
     if (room.owner !== user._id) {
-      throw new Forbidden();
+      throw new Forbidden(NOT_ROOM_OWNER_ERROR_MESSAGE);
     }
 
     const updatedRoom = await this.roomService.updateRoom({ ...room, name, slug, documentsMode, description });
@@ -120,7 +120,11 @@ export default class RoomController {
     }
 
     if (room.owner !== user._id) {
-      throw new Forbidden();
+      throw new Forbidden(NOT_ROOM_OWNER_ERROR_MESSAGE);
+    }
+
+    if (room.access === ROOM_ACCESS.public) {
+      throw new Forbidden('Public rooms cannot be deleted');
     }
 
     await this._deleteRoom({ room, roomOwner: user });
@@ -140,7 +144,7 @@ export default class RoomController {
     }
 
     if (room.owner !== user._id && memberUserId !== user._id) {
-      throw new Forbidden();
+      throw new Forbidden(NOT_ROOM_OWNER_OR_MEMBER_ERROR_MESSAGE);
     }
 
     const updatedRoom = await this.roomService.removeRoomMember({ room, memberUserId });
@@ -162,7 +166,7 @@ export default class RoomController {
 
     const room = await this.roomService.getRoomById(invitation.roomId);
     if (room.owner !== user._id) {
-      throw new Forbidden();
+      throw new Forbidden(NOT_ROOM_OWNER_ERROR_MESSAGE);
     }
 
     const remainingRoomInvitations = await this.roomService.deleteRoomInvitation({ room, invitation });
@@ -217,7 +221,7 @@ export default class RoomController {
     if (isPrivateRoom) {
       const isRoomOwnerOrMember = await this.roomService.isRoomOwnerOrMember(roomId, userId);
       if (!isRoomOwnerOrMember) {
-        throw new Forbidden();
+        throw new Forbidden(NOT_ROOM_OWNER_OR_MEMBER_ERROR_MESSAGE);
       }
     }
